Refactor rendering and texture management for bgfx integration
- Renamed font initialization methods and updated related private methods for clarity.
- Introduced stb_truetype for font handling, replacing GDI font dependencies.
- Updated renderer types to include new texture address and operation types.
- Refactored shader and sound object includes to use consistent casing.
- Replaced D3D8 dependencies in sorting renderer with Matrix4 math.
- Enhanced surface copy methods to utilize SurfaceClass for texture operations.
- Cleaned up texture loader and thumbnail manager to use modern C++ practices.
- Fixed include case issues across multiple files for Linux compatibility.
- Removed deprecated dependencies and streamlined texture management for bgfx.
- Added stb as a submodule for font rendering support.

+++ Code/ww3d2/part_ldr.cpp
@@ -1212,7 +1212,7 @@ ParticleEmitterDefClass::Save_User_Data (ChunkSaveClass &chunk_save)
 	// Begin a chunk that contains user information
 	if (chunk_save.Begin_Chunk (W3D_CHUNK_EMITTER_USER_DATA) == TRUE) {
 		
-		DWORD string_len = m_pUserString ? (::lstrlen (m_pUserString) + 1) : 0;
+		unsigned long string_len = m_pUserString ? (::lstrlen (m_pUserString) + 1) : 0;
 
 		// Fill the header structure
 		W3dEmitterUserInfoStruct user_info = { 0 };
